refactor du projet (suppress civ class) + continue batiment class
