+++ core/src/civ/game/Production.java
@@ -1,4 +1,4 @@
-package civ.game;
+package game;
 
 
 /**
