+++ core/src/civ/game/Ville.java
@@ -1,4 +1,4 @@
-package civ.game;
+package game;
 
 /**
  * Classe Ville permettant de generer les villes des joueurs et leurs actions associées.
