+++ core/src/civ/game/tuile/Contenu.java
@@ -1,6 +1,6 @@
-package civ.game.tuile;
+package game.tuile;
 
-import civ.game.Production;
+import game.Production;
 
 /**
  * Interface Contenu permettant de regrouper tout les types etant contenu dans une tuile.
@@ -21,8 +21,9 @@ public interface Contenu {
      */
     int getModificateurDeplacement();
 
-    /** 
+    /**
      * Permet d'obtenir la production du contenu de la tuile
+     *
      * @return Un objet Production resumant la production du contenu de la tuile
      */
     Production getProduction();
