+++ core/src/civ/game/tuile/contenu/Caracteristique.java
@@ -1,6 +1,6 @@
-package civ.game.tuile.contenu;
+package game.tuile.contenu;
 
-import civ.game.tuile.Contenu;
+import game.tuile.Contenu;
 
 public interface Caracteristique extends Contenu {
