+++ core/src/civ/game/tuile/contenu/caracteristique/ForetTropicale.java
@@ -1,7 +1,7 @@
-package civ.game.tuile.contenu.caracteristique;
+package game.tuile.contenu.caracteristique;
 
-import civ.game.Production;
-import civ.game.tuile.contenu.Caracteristique;
+import game.Production;
+import game.tuile.contenu.Caracteristique;
 
 public class ForetTropicale implements Caracteristique{
