+++ core/src/civ/game/tuile/terrain/Ocean.java
@@ -1,13 +1,13 @@
-package civ.game.tuile.terrain;
+package game.tuile.terrain;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import civ.game.Production;
-import civ.game.tuile.Terrain;
-import civ.game.tuile.contenu.Caracteristique;
+import game.Production;
+import game.tuile.Terrain;
+import game.tuile.contenu.Caracteristique;
 
-public class Ocean implements Terrain{
+public class Ocean implements Terrain {
     
     public int getModificateurCombat() {
         return 0;
