+++ desktop/src/civ/game/DesktopLauncher.java
@@ -2,6 +2,7 @@ package civ.game;
 
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
+import game.CivGame;
 
 
 // Please note that on macOS your application needs to be started with the -XstartOnFirstThread JVM argument
